+++ src/main/java/org/lxc/platform/respository/JobRepository.java
@@ -4,6 +4,6 @@ import java.util.Optional;
 import org.lxc.platform.model.Job;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface JobRepository extends JpaRepository<Job, Integer> {
+public interface JobRepository extends JpaRepository<Job, Long> {
   Optional<Job> findByKey(String key);
 }
